requirements.txt
httpx==0.28.1
openai==2.14.0
websockets==15.0.1
zai-sdk
dashscope
